refactor: controller 와 service 분리
