+++ src/routes/user/profile/index.ts
@@ -1,19 +1,14 @@
 import { Router } from 'express';
 import { checkToken } from '../../../middlewares/auth';
-import {
-  createProfile,
-  getUserProfileById,
-  updateUserProfileById,
-} from '../../../services/user/profile';
-import helloWorld from '../../hello-world';
+import controller from './profile.controller';
 
 const router = Router();
 
 // 프로필 설정 페이지
-router.post('/:id', createProfile);
+router.post('/:id', checkToken, controller.createProfile);
 // 사용자 프로필 정보 조회
-router.get('/:id', getUserProfileById);
+router.get('/:id', checkToken, controller.getUserProfileById);
 // 사용자 프로필 갱신
-router.patch('/:id', updateUserProfileById);
+router.patch('/:id', checkToken, controller.updateUserProfileById);
 
 export default router;
